+++ src/app/api/menu.ts
@@ -1,13 +1,6 @@
 import { instance } from '@/lib/axios';
 import { getStoreId } from './store';
-
-export type MenuItem = {
-  id: number;
-  name: string;
-  price: number;
-  category: string;
-  available: boolean;
-};
+import { MenuItem } from '@/types/menu';
 
 export async function getMenus(): Promise<MenuItem[]> {
   const storeId = getStoreId();
