+++ src/types/menu.ts
@@ -1,0 +1,7 @@
+export type MenuItem = {
+  id: number;
+  name: string;
+  price: number;
+  category: string;
+  available: boolean;
+};
